Initial commit
Include two images, one half float EXR and
one 16 bit TIFF. The TIFF is nonlinearly
encoded, the EXR linear. Both images are
encoded using a developer testing colorspace
which will reveal incorrect colour and
dynamic ranges if the input and output
transforms are not set correctly. When set
properly, the image colours and ranges will
align with the labels.